Refactor: Extract the OSX-specific code to a separate file
